updated my forecast appearance/ added javascript/rearranged the forecast div.
the styling was still not working like I would have liked  so i adjusted my fonts to vw so the width would change when the view changed. I think I like the search engine a lot now and can focus on the five-day forecast. I have rearranged the HTML of the weather forecast to work better with Flex and also for responsiveness.